+++ examples/next-openai/app/api/chat-openai-image-generation-custom-tool/route.ts
@@ -1,37 +1,12 @@
-import { openai } from '@ai-sdk/openai';
-import {
-  convertToModelMessages,
-  InferUITools,
-  streamText,
-  ToolSet,
-  UIDataTypes,
-  UIMessage,
-  validateUIMessages,
-} from 'ai';
-import { generateImageTool } from '@/tool/generate-image-tool';
+import { openaiImageGenerationCustomToolAgent } from '@/agent/openai-image-generation-custom-tool-agent';
+import { validateUIMessages } from 'ai';
 
-const tools = {
-  imageGeneration: generateImageTool,
-} satisfies ToolSet;
+export async function POST(request: Request) {
+  const body = await request.json();
 
-export type OpenAIImageGenerationMessage = UIMessage<
-  never,
-  UIDataTypes,
-  InferUITools<typeof tools>
->;
+  console.dir(body.messages, { depth: Infinity });
 
-export async function POST(req: Request) {
-  const { messages } = await req.json();
-  const uiMessages = await validateUIMessages({ messages });
-
-  const result = streamText({
-    model: openai('gpt-5-nano'),
-    tools,
-    messages: convertToModelMessages(uiMessages, { tools }),
-    onStepFinish: ({ request }) => {
-      console.dir(request.body, { depth: 3 });
-    },
+  return openaiImageGenerationCustomToolAgent.respond({
+    messages: await validateUIMessages({ messages: body.messages }),
   });
-
-  return result.toUIMessageStreamResponse();
 }