+++ examples/next-openai/app/chat-openai-pdf-fetch-custom-tool/page.tsx
@@ -1,22 +1,24 @@
 'use client';
 
-import { OpenAIImageGenerationMessage } from '@/app/api/chat-openai-image-generation-custom-tool/route';
 import ChatInput from '@/components/chat-input';
-import GenerateImageView from '@/components/tool/generate-image-view';
+import FetchPDFView from '@/components/tool/fetch-pdf-view';
 import { useChat } from '@ai-sdk/react';
 import { DefaultChatTransport } from 'ai';
+import { OpenAIFetchPdfCustomToolMessage } from '@/agent/openai-fetch-pdf-custom-tool-agent';
 
-export default function TestOpenAIWebSearch() {
+export default function TestOpenAIFileGenerationPDF() {
   const { status, sendMessage, messages } =
-    useChat<OpenAIImageGenerationMessage>({
+    useChat<OpenAIFetchPdfCustomToolMessage>({
       transport: new DefaultChatTransport({
-        api: '/api/chat-openai-image-generation-custom-tool',
+        api: '/api/chat-openai-fetch-pdf-custom-tool',
       }),
     });
 
   return (
     <div className="flex flex-col py-24 mx-auto w-full max-w-md stretch">
-      <h1 className="mb-4 text-xl font-bold">OpenAI Image Generation Test</h1>
+      <h1 className="mb-4 text-xl font-bold">
+        OpenAI Fetch PDF Custom Tool Test
+      </h1>
 
       {messages.map(message => (
         <div key={message.id} className="whitespace-pre-wrap">
@@ -25,8 +27,8 @@ export default function TestOpenAIWebSearch() {
             switch (part.type) {
               case 'text':
                 return <div key={index}>{part.text}</div>;
-              case 'tool-imageGeneration':
-                return <GenerateImageView key={index} invocation={part} />;
+              case 'tool-fetchPdf':
+                return <FetchPDFView key={index} invocation={part} />;
             }
           })}
         </div>