+++ packages/ai/src/ui/validate-ui-messages.ts
@@ -80,8 +80,10 @@ const uiMessagesSchema = lazySchema(() =>
               toolCallId: z.string(),
               state: z.literal('input-streaming'),
               input: z.unknown().optional(),
+              providerExecuted: z.boolean().optional(),
               output: z.never().optional(),
               errorText: z.never().optional(),
+              approval: z.never().optional(),
             }),
             z.object({
               type: z.literal('dynamic-tool'),
@@ -89,9 +91,43 @@ const uiMessagesSchema = lazySchema(() =>
               toolCallId: z.string(),
               state: z.literal('input-available'),
               input: z.unknown(),
+              providerExecuted: z.boolean().optional(),
               output: z.never().optional(),
               errorText: z.never().optional(),
               callProviderMetadata: providerMetadataSchema.optional(),
+              approval: z.never().optional(),
+            }),
+            z.object({
+              type: z.literal('dynamic-tool'),
+              toolName: z.string(),
+              toolCallId: z.string(),
+              state: z.literal('approval-requested'),
+              input: z.unknown(),
+              providerExecuted: z.boolean().optional(),
+              output: z.never().optional(),
+              errorText: z.never().optional(),
+              callProviderMetadata: providerMetadataSchema.optional(),
+              approval: z.object({
+                id: z.string(),
+                approved: z.never().optional(),
+                reason: z.never().optional(),
+              }),
+            }),
+            z.object({
+              type: z.literal('dynamic-tool'),
+              toolName: z.string(),
+              toolCallId: z.string(),
+              state: z.literal('approval-responded'),
+              input: z.unknown(),
+              providerExecuted: z.boolean().optional(),
+              output: z.never().optional(),
+              errorText: z.never().optional(),
+              callProviderMetadata: providerMetadataSchema.optional(),
+              approval: z.object({
+                id: z.string(),
+                approved: z.boolean(),
+                reason: z.string().optional(),
+              }),
             }),
             z.object({
               type: z.literal('dynamic-tool'),
@@ -99,10 +135,18 @@ const uiMessagesSchema = lazySchema(() =>
               toolCallId: z.string(),
               state: z.literal('output-available'),
               input: z.unknown(),
+              providerExecuted: z.boolean().optional(),
               output: z.unknown(),
               errorText: z.never().optional(),
               callProviderMetadata: providerMetadataSchema.optional(),
               preliminary: z.boolean().optional(),
+              approval: z
+                .object({
+                  id: z.string(),
+                  approved: z.literal(true),
+                  reason: z.string().optional(),
+                })
+                .optional(),
             }),
             z.object({
               type: z.literal('dynamic-tool'),
@@ -110,9 +154,33 @@ const uiMessagesSchema = lazySchema(() =>
               toolCallId: z.string(),
               state: z.literal('output-error'),
               input: z.unknown(),
+              providerExecuted: z.boolean().optional(),
               output: z.never().optional(),
               errorText: z.string(),
               callProviderMetadata: providerMetadataSchema.optional(),
+              approval: z
+                .object({
+                  id: z.string(),
+                  approved: z.literal(true),
+                  reason: z.string().optional(),
+                })
+                .optional(),
+            }),
+            z.object({
+              type: z.literal('dynamic-tool'),
+              toolName: z.string(),
+              toolCallId: z.string(),
+              state: z.literal('output-denied'),
+              input: z.unknown(),
+              providerExecuted: z.boolean().optional(),
+              output: z.never().optional(),
+              errorText: z.never().optional(),
+              callProviderMetadata: providerMetadataSchema.optional(),
+              approval: z.object({
+                id: z.string(),
+                approved: z.literal(false),
+                reason: z.string().optional(),
+              }),
             }),
             z.object({
               type: z.string().startsWith('tool-'),
@@ -321,6 +389,7 @@ export async function safeValidateUIMessages<UI_MESSAGE extends UIMessage>({
           const toolName = toolPart.type.slice(5);
           const tool = tools[toolName];
 
+          // TODO support dynamic tools
           if (!tool) {
             return {
               success: false,